requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Migrate>=3.1.0
Flask-Login>=0.5.0
Flask-CORS>=3.0.10
Flask-Session>=0.4.0
Flask-Babel>=2.0.0
Flask-Admin>=1.6.0
Flask-WTF>=1.0.0
WTForms>=3.0.0
python-dotenv>=0.19.0
deep-translator>=1.11.0
beautifulsoup4>=4.9.1
requests>=2.23.0
python-slugify>=5.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0